Reorganize some storage code.
