+++ src/FreeDSx/Ldap/Server/Backend/Storage/Filter/FilterEvaluatorInterface.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace FreeDSx\Ldap\Server\Backend\Storage;
+namespace FreeDSx\Ldap\Server\Backend\Storage\Filter;
 
 use FreeDSx\Ldap\Entry\Entry;
 use FreeDSx\Ldap\Search\Filter\FilterInterface;
